Initial attempt on adding weights information
I added an option to pass in c1 and c2 via `weight_params` argument in
Datassset and Configuration. The weights are calculated by $1/w = c1^2
+ c2^2 * \Vert f \Vert^2$. The property `config.weight` is a
dictionary containing the computed weights in this way. I modified the
default residual function to use these weights.

I haven't checked any of the tests.