Create class that handles weight
This is Mingjian's suggestion to have a class for weight. The base
class is hoping to replace the weight keys in the `residual_data`. I
also implement another weight class that computes nonuniform weight,
as mention previously.
There are also some changes in loss.py, mainly to just revert it back
and delete the weight keys in `residual_data`.
The test_scipy_optimize.py is also updated to use this convention.