Fix bug if the weight needs configuration's identifier
This bug occurs when the weight value depends on the configuration's
identifier. For example, we might want to weigh the configurations
with cracks. The
[documentation](https://kliff.readthedocs.io/en/latest/modules/dataset.html#define-your-weight-class)
shows an example on how this can be done, which uses the identifier to
tell the function whether the configuration has cracks.

Previously, I didn't test this thoroughly and apparently we will get
an error if we apply this example, since the identifier is defined
after `weight.compute_weight()` is called.

I pretty much just swapped the order: now we call
`weight.compute_weight()` after the identifier is defined.