Inherit from ptemcee and emcee samplers
In the previous version, we instantiate `ptemcee.Sampler` or
`emcee.EnsembleSampler` when we instantiate the MCMC class, i.e.,
we call `self.sampler = ptemcee.Sampler()`. I noticed that doing
this somehow caused problem when we try hybrid parallelization,
for example, using multithreading in the loss evaluation and MPI
in the sampling.
This issue seems to be solved when we instead inherit the MCMC class
from ptemcee or emcee.
The performance of hybrid parallelization using for ptemcee sampling
has been tested, but the emcee sampling case is still yet to be tested.