Fix bug about defining sampler based on whether package are found
In this version, we will not throw an import error if ptemcee/emcee is
not installed before defining the corresponding class. Otherwise, I
believe that if we do like in the previous version, it will always
throw an import error if we try to import `kliff.uq.mcmc` when
ptemcee/emcee is not installed.