Cache the initial parameter guess for empirical model
We want to cache the initial guess that we set prior to training
because in the bootstrap, we want to start the training NOT from the
last optimal parameter. If we use the last optimal values, the result
might be biased from seeing the entire dataset. We want to treat each
step in the bootstrap sampling as though the bootstrap sample compute
arguments are the only data we have.