Reset the parameters for NN and empirical model
For NN model, we use `reset_parameters` method implemented in each
layer module.
For the empirical model, I removed the option to have custom function
to get the initial guess. This is done to make the class behave more
similarly to the bootstrap NN class. However, this can be implemented
in the future, if needed.