Fix the list of residual function when using _WrapperCalculator
When the default bootstrap cas generator is used with
_WrapperCalculator, the number of configurations for each calculator
might be different than the original list. As such, the old list of
residual functions might not be appropriate for some configurations.
For example if we use the original list of residual functions, we might
use `forces_residual` when the bootstrap configuration only compute
for energy.