Add additional argument for default bootstrap dataset generator
I added an argument to specify how many bootstrap compute arguments to
generate for each sample. Generally, in each sample we want to have
the same number of compute arguments as the number of compute
arguments in the original list. However, with this option there might
be an interesting study that can be done, e.g., if we use fewer
compute arguments in each sample.