Update how to compute sigma in MagnitudeInverseWeight
There was an issue that the old way to compute sigma raised a
`VisibleDeprecationWarning` for some python version and error for
Python > 3.9. I think this was because we were using list of float and
array inside `np.linalg.norm` to compute the norm of value
corresponding to each data point.