Fix how ptemcee dependency is listed in `extra_requires`
We use Yonatan's fork of ptemcee as dependency, but that version of
package is not registered in pypi. So, we need to specify the link to
the repo (and the specific branch). But, apparently there is a
specific way how we can list such repo. So, this commit fixes this problem.