Fix bug when we try to set NN parameter with different dtype
torch defaults to use float32. But, maybe we want to update the
weights and biases in the NN model using a parameter array written as
a numpy array, which defaults to usin float64. This commit fixes this
problem and convert the parameter array to use whatever dtype used by
the NN model.